refactor(langauge_server): move `Backend` into own file (#13955)
moved the `Backend` struct into an own file so `main.rs` includes only the "startup".
Created a `Backend::new` method for simplifying the creation
