+++ crates/oxc_language_server/src/worker.rs
@@ -11,7 +11,7 @@ use tower_lsp_server::{
 };
 
 use crate::{
-    ConcurrentHashMap, Options,
+    ConcurrentHashMap,
     code_actions::{
         apply_all_fix_code_action, apply_fix_code_actions, ignore_this_line_code_action,
         ignore_this_rule_code_action,
@@ -21,6 +21,7 @@ use crate::{
         error_with_position::{DiagnosticReport, PossibleFixContent},
         server_linter::{ServerLinter, ServerLinterRun, normalize_path},
     },
+    options::Options,
 };
 
 pub struct WorkspaceWorker {
